replaced redux for cities with react query and refactored some code in navbar
