+++ src/features/authentication/utils/getEmail.ts
@@ -1,0 +1,3 @@
+import { getItem } from '../../../utils/localStorage';
+
+export const getEmail = (): string | undefined => getItem('email');
