+++ src/features/authentication/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './getEmail';
+export * from './isAuthenticated';
+export * from './logout';
