+++ src/features/authentication/utils/isAuthenticated.ts
@@ -1,0 +1,6 @@
+import { getItem } from '../../../utils/localStorage';
+
+export const isAuthenticated = (): boolean => {
+  const token = getItem('token');
+  return token ? true : false;
+};
