+++ src/features/authentication/utils/logout.ts
@@ -1,0 +1,9 @@
+import { getItem, setItem } from '../../../utils/localStorage';
+
+export const logout = (): void => {
+  const city = getItem('city');
+  localStorage.clear();
+  setItem('city', city);
+
+  location.href = '/';
+};
