+++ src/features/movies/RecommendedMovieItem.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 import { FaStar } from 'react-icons/fa';
 
-import { Releases } from '../../services/movies/apiReleases';
+import { Releases } from './services/apiReleases';
 
 import createSlug from '../../utils/createSlug';
