+++ src/store.ts
@@ -1,13 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import userReducer from './features/authentication/slices/authSlice';
-import citiesReducer from './features/cities/citiesSlice';
+// import userReducer from './features/authentication/slices/authSlice';
+// import citiesReducer from './features/cities/citiesSlice';
 // import releasesReducer from './features/movies/moviesSlice';
 
 const store = configureStore({
   reducer: {
-    auth: userReducer,
-    cities: citiesReducer,
+    // auth: userReducer
+    // cities: citiesReducer,
     // releases: releasesReducer
   }
 });
