+++ src/ui/Hamburger.tsx
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 
 import HamburgerMenuList from './HamburgerMenuList';
 import { IoChevronForward } from 'react-icons/io5';
-
-import { getEmail, isAuthenticated, logout } from '../utils/auth';
+import { getEmail, isAuthenticated, logout } from '../features/authentication/utils';
 
 interface HamburgerMenuProps {
   onShowSignIn(): void;
