--- src/utils/getCity.ts
@@ -1,6 +1,0 @@
-const getCity = (): string => {
-  const city = localStorage.getItem('city');
-  return city || '';
-};
-
-export default getCity;
